+++ lib/schleuder/keyword_handlers/subscribe.rb
@@ -2,8 +2,8 @@ module Schleuder
   module KeywordHandlers
     class Subscribe < Base
       # This class parses its keyword arguments itself, but the method argument keyword must be given nonetheless.
-      handles_request_keyword :subscribe, with_arguments: []
-      handles_request_keyword 'add-member', with_arguments: [] # Example of an alias
+      handles_request_keyword :subscribe, with_arguments: [:emailaddr, :fingerprint, :bool, :bool]
+      handles_request_keyword 'add-member', with_arguments: [:emailaddr, :fingerprint, :bool, :bool] # Example of an alias
 
       # This method returns the collected arguments as second return value so
       # the actual method doesn't have to check for a space-delimited